Make playground use the card store
Updated playground to use the card store for none web cards

Deleted all connection profile store related code

closes to hyperledger/composer#2275

+++ packages/composer-playground/src/app/services/samplebusinessnetwork.service.ts
@@ -6,7 +6,7 @@ import { ClientService } from './client.service';
 import { FileService } from './file.service';
 import { AlertService } from '../basic-modals/alert.service';
 
-import { BusinessNetworkDefinition } from 'composer-common';
+import { BusinessNetworkDefinition, IdCard } from 'composer-common';
 import { IdentityCardService } from './identity-card.service';
 
 @Injectable()
@@ -58,9 +58,8 @@ export class SampleBusinessNetworkService {
             });
     }
 
-    generateBootstrapTransactions(businessNetworkDefinition: BusinessNetworkDefinition, identityName: string, credentials): Object[] {
+    generateAddParticpantTransaction(identityName: string, businessNetworkDefinition: BusinessNetworkDefinition): string {
         const factory = businessNetworkDefinition.getFactory();
-        const serializer = businessNetworkDefinition.getSerializer();
         const participant = factory.newResource('org.hyperledger.composer.system', 'NetworkAdmin', identityName);
         const targetRegistry = factory.newRelationship('org.hyperledger.composer.system', 'ParticipantRegistry', participant.getFullyQualifiedType());
         const addParticipantTransaction = factory.newTransaction('org.hyperledger.composer.system', 'AddParticipant');
@@ -68,6 +67,15 @@ export class SampleBusinessNetworkService {
             resources: [participant],
             targetRegistry
         });
+
+        return addParticipantTransaction;
+    }
+
+    generateBootstrapTransactions(businessNetworkDefinition: BusinessNetworkDefinition, identityName: string, credentials): Object[] {
+        const factory = businessNetworkDefinition.getFactory();
+        const serializer = businessNetworkDefinition.getSerializer();
+
+        const addParticipantTransaction = this.generateAddParticpantTransaction(identityName, businessNetworkDefinition);
 
         let identityTransaction;
 
@@ -96,22 +104,44 @@ export class SampleBusinessNetworkService {
         return result;
     }
 
-    public deployBusinessNetwork(businessNetworkDefinition: BusinessNetworkDefinition, networkName: string, networkDescription: string, networkId: string, networkSecret: string, credentials): Promise<string> {
+    public deployBusinessNetwork(businessNetworkDefinition: BusinessNetworkDefinition, cardName: string, networkName: string, networkDescription: string, networkId: string, networkSecret: string, credentials): Promise<string> {
         let packageJson = businessNetworkDefinition.getMetadata().getPackageJson();
         packageJson.name = networkName;
         packageJson.description = networkDescription;
 
         let newCardRef: string;
         let channelAdminCardRef: string;
+        let channelAdminCard: IdCard;
+        let peerAdminCardRef: string = this.identityCardService.getCurrentCardRef();
+        let peerAdminCard: IdCard = this.identityCardService.getIdentityCard(peerAdminCardRef);
 
         let newNetwork = this.buildNetwork(networkName, networkDescription, packageJson, businessNetworkDefinition);
 
-        // we should already be the PeerAdmin at this point, but just to check
-        let roles = this.identityCardService.getCurrentIdentityCard().getRoles();
-        if (!roles.includes('PeerAdmin')) {
-            return Promise.reject('The current identity does not have the role PeerAdmin, this role is required to install the business network');
-        }
-        return this.adminService.connectWithoutNetwork(true)
+        return Promise.resolve()
+            .then(() => {
+                if (networkSecret) {
+                    return this.identityCardService.createIdentityCard(networkId, cardName, newNetwork.getName(), networkSecret, peerAdminCard.getConnectionProfile())
+                        .then((idCardRef: string) => {
+                            newCardRef = idCardRef;
+                        });
+                } else if (credentials) {
+                    return this.identityCardService.createIdentityCard(networkId, cardName, newNetwork.getName(), null, peerAdminCard.getConnectionProfile(), credentials)
+                        .then((idCardRef: string) => {
+                            newCardRef = idCardRef;
+                        });
+                } else {
+                    networkId = 'admin';
+                    return this.identityCardService.createIdentityCard(networkId, cardName, newNetwork.getName(), 'adminpw', peerAdminCard.getConnectionProfile())
+                        .then((idCardRef: string) => {
+                            newCardRef = idCardRef;
+                        });
+                }
+            })
+            .then(() => {
+                let card = this.identityCardService.getIdentityCard(peerAdminCardRef);
+                console.log('peerCard', card);
+                return this.adminService.connect(peerAdminCardRef, card, true);
+            })
             .then(() => {
                 this.alertService.busyStatus$.next({
                     title: 'Installing Business Network',
@@ -120,50 +150,12 @@ export class SampleBusinessNetworkService {
                 return this.adminService.install(newNetwork.getName());
             })
             .then(() => {
-                let connectionProfile = this.identityCardService.getCurrentIdentityCard().getConnectionProfile();
+                let connectionProfile = this.identityCardService.getIdentityCard(peerAdminCardRef).getConnectionProfile();
                 let qpn = this.identityCardService.getQualifiedProfileName(connectionProfile);
                 channelAdminCardRef = this.identityCardService.getIdentityCardRefsWithProfileAndRole(qpn, 'ChannelAdmin')[0];
 
-                return this.identityCardService.setCurrentIdentityCard(channelAdminCardRef);
-            })
-            .then(() => {
-                return this.adminService.connectWithoutNetwork(true);
-            })
-            .then(() => {
-                if (networkSecret) {
-                    return this.identityCardService.createIdentityCard(networkId, newNetwork.getName(), networkSecret, this.identityCardService.getCurrentIdentityCard().getConnectionProfile())
-                        .then((idCardRef: string) => {
-                            newCardRef = idCardRef;
-                            return this.identityCardService.setCurrentIdentityCard(idCardRef);
-                        })
-                        .then(() => {
-                            return this.adminService.connectWithoutNetwork(true);
-                        })
-                        .then(() => {
-                            let connectionProfile = this.identityCardService.getCurrentIdentityCard().getConnectionProfile();
-                            let qpn = this.identityCardService.getQualifiedProfileName(connectionProfile);
-                            return this.adminService.exportIdentity(qpn, networkId);
-                        })
-                        .then((retrievedCertificate) => {
-                            credentials = {};
-                            credentials.certificate = retrievedCertificate.certificate;
-                            return this.identityCardService.setCurrentIdentityCard(channelAdminCardRef);
-                        })
-                        .then(() => {
-                            return this.adminService.connectWithoutNetwork(true);
-                        });
-                } else if (credentials) {
-                    return this.identityCardService.createIdentityCard(networkId, newNetwork.getName(), null, this.identityCardService.getCurrentIdentityCard().getConnectionProfile(), credentials)
-                        .then((idCardRef: string) => {
-                            newCardRef = idCardRef;
-                        });
-                } else {
-                    networkId = 'admin';
-                    return this.identityCardService.createIdentityCard(networkId, newNetwork.getName(), 'adminpw', this.identityCardService.getCurrentIdentityCard().getConnectionProfile())
-                        .then((idCardRef: string) => {
-                            newCardRef = idCardRef;
-                        });
-                }
+                channelAdminCard = this.identityCardService.getIdentityCard(channelAdminCardRef);
+                return this.adminService.connect(channelAdminCardRef, channelAdminCard, true);
             })
             .then(() => {
                 this.alertService.busyStatus$.next({
@@ -173,14 +165,14 @@ export class SampleBusinessNetworkService {
 
                 const bootstrapTransactions = this.generateBootstrapTransactions(businessNetworkDefinition, networkId, credentials);
 
-                return this.adminService.start(newNetwork, {bootstrapTransactions});
+                return this.adminService.start(newNetwork, {bootstrapTransactions, card : channelAdminCard});
             })
             .then(() => {
                 this.alertService.busyStatus$.next(null);
                 return newCardRef;
             })
             .catch((error) => {
-                if (newCardRef) {
+                if (newCardRef && !error.message.includes('already exists')) {
                     return this.identityCardService.deleteIdentityCard(newCardRef)
                         .then(() => {
                             this.alertService.busyStatus$.next(null);
@@ -208,7 +200,10 @@ export class SampleBusinessNetworkService {
 
         let newNetwork = this.buildNetwork(currentBusinessNetworkName, currentBusinessNetworkDescription, packageJson, businessNetworkDefinition);
 
-        return this.adminService.connect(currentBusinessNetworkName, true)
+        let cardName = this.identityCardService.getCurrentCardRef();
+        let card = this.identityCardService.getCurrentIdentityCard();
+
+        return this.adminService.connect(cardName, card, true)
             .then(() => {
                 this.alertService.busyStatus$.next({
                     title: 'Updating business network'
@@ -217,7 +212,7 @@ export class SampleBusinessNetworkService {
                 return this.adminService.update(newNetwork);
             })
             .then(() => {
-                return this.clientService.refresh(newNetwork.getName());
+                return this.clientService.refresh();
             })
             .then(() => {
                 return this.adminService.reset(newNetwork.getName());
